refactor: Changes made to file names for functionality understanding
