+++ pigz-app/src/App.js
@@ -1,14 +1,14 @@
 import './App.css';
 import Header from './Components/Header';
-import MainScreen from './Page/MainScreen';
+import RegistrationContent from './Page/RegistrationContent';
 import PigzLogo from './Components/PigzLogo'
-import AboutApp from './Page/AboutApp';
+import AboutApp from './Page/FeaturesContent';
 
 function App() {
   return (
     <div className="App">
       <Header />
-      <MainScreen />
+      <RegistrationContent />
       <PigzLogo />
       <AboutApp />
     </div>
