--- pigz-app/src/Page/MainScreen.js
@@ -1,11 +1,0 @@
-import RegistrationForm from "../Components/RegistrationForm";
-
-function MainScreen() {
-    return(
-        <div>
-            <RegistrationForm />
-        </div>
-    )
-}
-
-export default MainScreen;
