+++ pigz-app/src/Page/RegistrationContent.js
@@ -1,0 +1,14 @@
+import RegistrationTextCard from "../Components/RegistrationTextCard";
+import RegistrationFormCard from "../Components/RegistrationFormCard";
+
+
+function RegistrationContent() {
+    return(
+        <div>
+            <RegistrationTextCard />
+            <RegistrationFormCard />
+        </div>
+    )
+}
+
+export default RegistrationContent;
