+++ pigz-app/src/Page/SupportContent.js
@@ -1,0 +1,10 @@
+function SupportPigz() {
+    return (
+      <div>
+        <h1>Pigz Marketplace</h1>
+      </div>
+    );
+  }
+  
+  export default SupportPigz;
+
